Add basic Connection Pool ability
Currently supports PostgreSQL only. Other drivers like MySQL or Oracle need to accept dispose-callback and add a scrub-state-for-reuse() function which resets the connection state. I don't believe there would be a benefit for SQLite.

This is implemented as a Channel for the queue of connections with a few atomic counters to assist in counting the number of active database connections and their state. The connection itself is exactly what you would get from DBIish.connect() aside from the addition of a dispose-callback hook.

The pool has an initial size, an idle timeout, and enforces max-connections. At max-connections $pool.connect will wait until a currently in-use connection becomes available. A max work-load is typically useful to prevent overloading the database.

Typical use looks like this, where the connection arguments are passed to the driver specified.

	my $pool = DBIish.pool('Pg', max-connections => 10, |%conn);
    my $dbh = $pool.connect()
    # DO WORK
    $dbh.dispose;

For Pg, execution time for 20_000 connect/SELECT 1/disconnect loops is reduced from 23 seconds to 5 seconds for a database on localhost. This difference would be quite a bit higher for a networked database.